Fixed data types in Create_CL_MemBlock()

+++ src/clHeap.cpp
@@ -1,7 +1,7 @@
 #include "clHeap.hpp"
 #include <iostream>
 
-void Create_CL_MemBlock(CL_MemBlock *memBlock, uint32_t start, uint32_t end, uint64_t size, bool free) {
+void Create_CL_MemBlock(CL_MemBlock *memBlock, uint64_t start, uint64_t end, uint64_t size, bool free) {
     memBlock->start = start;
     memBlock->end = end;
     memBlock->size = size;
